refactor(tui): extract TasksBrowserController from KimiTUI
将 /tasks 浏览器相关的 16 个方法（~450 行）从 KimiTUI 提取到独立的
TasksBrowserController。KimiTUI 通过 Host 接口传 this 给 controller，
保留 4 个一行委托方法供外部调用点使用。

kimi-tui.ts: 6160 → 5711 行 (-449)
